Refactor: split utils module
