+++ gn_django/utils/models.py
@@ -1,0 +1,22 @@
+from django.utils.text import slugify
+
+
+def unique_object_slug(obj, source, slug_field='slug', limit=10000):
+    """
+    Utility for generating unique slugs based on a source string. For instance
+    converting an article title of 'Bloodborne 2 Announced' to 'bloodborne-2-announced'.
+    If the slug is already taken it will append a numeric value, i.e. 'bloodborne-2-announced-1'
+    """
+    model = obj.__class__
+    base_slug = slugify(source)
+    current_slug = base_slug
+    search = '%s__startswith' % slug_field
+    objects = model.objects.filter(**{search: base_slug})
+
+    for i in range(1, limit):
+        o = objects.filter(**{slug_field: current_slug})
+        if (len(o) == 0) or (o[0].id == obj.id):
+            return current_slug
+        current_slug = '%s-%s' % (base_slug, i)
+
+    return False
